Add libmraa GPIO I/O support
When libmraa is available, build a set of I/O
functions that make use of it rather than x86
parallel port operations.  This should work on
any platform that supports GPIOs, including x86,
so allow selecting between the x86 parallel port
and GPIO  paths at runtime if both are available.