Add '-d' option to upload raw data w/2005 loader
The 2005 version of the BJL loader added an
alternate magic word (20102001) that indicates to
the loader that data, rather than a program, is
being uploaded.  When this option is used, the
loader receives the data and places it at the
requested location in the Jaguar's RAM as usual,
but it does not attempt to execute it afterwards,
allowing more data or a program to be uploaded to
another location afterwards.

Example usage, testing cinepak playback from RAM:

  # Upload track synchronization header data
  lo_inp -8 -b 0x3ffc0 -d sync
  # Upload Star Wars escape sequence clip
  lo_inp -n -8 -b 0x40000 -d esc6.crg
  # Upload the sample player and execute it.
  lo_inp -n -8 player.cof